+++ cligen_callback.h
@@ -49,6 +49,9 @@ typedef int (cgv_fnstype_t)(cligen_handle h, cvec *cvv, cvec *argv);
 
 typedef struct cg_callback cg_callback; /* struct defined in cligen_callback.c.c */
 
+/* Callback flags */
+#define CC_FLAGS_PIPE_FUNCTION 0x01 /* Pipe cligen_output through callback function */
+
 /*! A CLIgen object may have one or several callbacks. This type defines one callback
  */
 struct cg_callback  { /* Linked list of command callbacks */
@@ -56,6 +59,7 @@ struct cg_callback  { /* Linked list of command callbacks */
     cgv_fnstype_t       *cc_fn_vec;  /**< callback/function pointer using cvec.  */
     char                *cc_fn_str;  /**< callback/function name. malloced */
     cvec                *cc_cvec;    /**< callback/function arguments */
+    int                  cc_flags;   /**< Callback flags, see CC_FLAGS_* above */
 };
 
 /*